+++ project-base/app/Environment.php
@@ -4,16 +4,15 @@ namespace Shopsys;
 
 use Composer\IO\IOInterface;
 use Composer\Script\Event;
+use Shopsys\FrameworkBundle\Component\Environment\EnvironmentFileSetting;
+use Shopsys\FrameworkBundle\Component\Environment\EnvironmentType;
 
 class Environment
 {
-    const ENVIRONMENT_PRODUCTION = 'prod';
-    const ENVIRONMENT_DEVELOPMENT = 'dev';
-    const ENVIRONMENT_TEST = 'test';
-
-    const FILE_DEVELOPMENT = 'DEVELOPMENT';
-    const FILE_PRODUCTION = 'PRODUCTION';
-    const FILE_TEST = 'TEST';
+    /**
+     * @var \Shopsys\FrameworkBundle\Component\Environment\EnvironmentFileSetting|null
+     */
+    private static $environmentFileSetting;
 
     /**
      * @param \Composer\Script\Event $event
@@ -22,11 +21,12 @@ class Environment
     {
         $io = $event->getIO();
         /* @var $io \Composer\IO\IOInterface */
-        if ($io->isInteractive() && self::getEnvironmentSetting(false) === null) {
+        $environmentFileSetting = self::getEnvironmentFileSetting();
+        if ($io->isInteractive() && !$environmentFileSetting->isAnyEnvironmentSet()) {
             if ($io->askConfirmation('Build in production environment? (Y/n): ', true)) {
-                self::createFile(self::getRootDir() . '/' . self::FILE_PRODUCTION);
+                $environmentFileSetting->createFileForEnvironment(EnvironmentType::PRODUCTION);
             } else {
-                self::createFile(self::getRootDir() . '/' . self::FILE_DEVELOPMENT);
+                $environmentFileSetting->createFileForEnvironment(EnvironmentType::DEVELOPMENT);
             }
         }
         self::printEnvironmentInfo($io);
@@ -38,17 +38,9 @@ class Environment
      */
     public static function getEnvironment($console)
     {
-        $environmentSetting = self::getEnvironmentSetting($console);
-        return $environmentSetting ?: self::ENVIRONMENT_PRODUCTION;
+        return self::getEnvironmentFileSetting()->getEnvironment($console);
     }
 
-    /**
-     * @param string $environment
-     */
-    public static function isEnvironmentDebug($environment)
-    {
-        return $environment === self::ENVIRONMENT_DEVELOPMENT;
-    }
 
     /**
      * @param \Composer\IO\IOInterface $io
@@ -59,35 +51,13 @@ class Environment
     }
 
     /**
-     * @param string $filepath
+     * @return \Shopsys\FrameworkBundle\Component\Environment\EnvironmentFileSetting
      */
-    private static function createFile($filepath)
+    private static function getEnvironmentFileSetting()
     {
-        $file = fopen($filepath, 'w');
-        fclose($file);
-    }
-
-    /**
-     * @return string
-     */
-    private static function getRootDir()
-    {
-        return __DIR__ . '/..';
-    }
-
-    /**
-     * @param bool $ignoreTestFile
-     * @return string|null
-     */
-    private static function getEnvironmentSetting($ignoreTestFile)
-    {
-        if (!$ignoreTestFile && is_file(self::getRootDir() . '/' . self::FILE_TEST)) {
-            return self::ENVIRONMENT_TEST;
-        } elseif (is_file(self::getRootDir() . '/' . self::FILE_DEVELOPMENT)) {
-            return self::ENVIRONMENT_DEVELOPMENT;
-        } elseif (is_file(self::getRootDir() . '/' . self::FILE_PRODUCTION)) {
-            return self::ENVIRONMENT_PRODUCTION;
+        if (self::$environmentFileSetting === null) {
+            self::$environmentFileSetting = new EnvironmentFileSetting(__DIR__ . '/..');
         }
-        return null;
+        return self::$environmentFileSetting;
     }
 }